+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.3.5 (unreleased)
+
+- Add `sdwan_security_policy` resource and data source
+
 ## 0.3.4
 
 - Add `sdwan_switchport_feature_template` resource and data source